rename grabbids to layout, closes #228
