+++ bids/__init__.py
@@ -3,12 +3,14 @@ from .due import due, Doi
 
 from .grabbids import BIDSLayout, BIDSValidator
 from .analysis import Analysis
+from .layout import BIDSLayout, BIDSValidator
 from .variables.io import load_variables
 
 
 __all__ = [
     "grabbids",
     "analysis",
+    "layout",
     "reports",
     "BIDSLayout",
     "BIDSValidator",
