+++ bids/layout/config/bids.json
