+++ bids/layout/tests/test_validator.py
@@ -2,7 +2,7 @@
 functionality should go in the grabbit package. """
 
 import pytest
-from bids.grabbids import BIDSValidator
+from bids.layout import BIDSValidator
 from os.path import join, dirname, abspath
