+++ bids/reports/tests/test_report.py
@@ -4,7 +4,7 @@ from os.path import join, abspath
 
 import pytest
 
-from bids.grabbids import BIDSLayout
+from bids.layout import BIDSLayout
 from bids.reports import BIDSReport
 from bids.tests import get_test_data_path
