+++ bids/variables/tests/test_io.py
@@ -1,4 +1,4 @@
-from bids.grabbids import BIDSLayout
+from bids.layout import BIDSLayout
 from bids.variables import (SparseRunVariable, SimpleVariable,
                             DenseRunVariable, load_variables)
 from bids.variables.entities import Node, RunNode, NodeIndex
